requirements.txt
joblib
matplotlib
numpy
pandas
python-dotenv
pytest
requests
scikit-learn
scipy
torch
xgboost
yfinance
